+++ CHANGELOG.md
@@ -18,6 +18,7 @@
  - Added memory usage statistics to request (#89)
  - Suppress backtraces for Controller, Job, and Middleware spans (#96)
  - Tag the request URI automatically with `$_SERVER['REQUEST_URI']` or override with alternative (#102)
+ - Update CoreAgent to 1.2.4 (#103)
 
 ### Fixed